refactor(bot): use more flexible interaction classes
