+++ apps/bot/package.json
@@ -17,8 +17,8 @@
 		"@ticketer/djs-framework": "workspace:*",
 		"@ticketer/env": "workspace:*",
 		"chalk": "^5.3.0",
-		"discord.js": "^14.16.1",
-		"tsx": "^4.19.0",
+		"discord.js": "^14.16.2",
+		"tsx": "^4.19.1",
 		"typesafe-i18n": "^5.26.2",
 		"zod": "^3.23.8"
 	},
