+++ apps/bot/src/commands/automatic-threads/lock-and-close.ts
@@ -1,7 +1,7 @@
 import { Component, DeferReply } from '@ticketer/djs-framework';
 import { AutomaticThreads } from '@/utils';
 
-export class ComponentInteraction extends Component.Interaction {
+export default class extends Component.Interaction {
 	public readonly customIds = [super.customId('ticket_automatic_threads_thread_lock_and_close')];
 
 	@DeferReply({ ephemeral: true })
