+++ apps/bot/src/commands/user-forums/lock.ts
@@ -1,7 +1,7 @@
 import { Component, DeferReply } from '@ticketer/djs-framework';
 import { UserForums } from '@/utils';
 
-export class ComponentInteraction extends Component.Interaction {
+export default class ComponentInteraction extends Component.Interaction {
 	public readonly customIds = [super.customId('ticket_user_forums_thread_lock')];
 
 	@DeferReply({ ephemeral: true })
