+++ apps/website/package.json
@@ -20,8 +20,8 @@
 		"@vercel/speed-insights": "^1.0.12",
 		"class-variance-authority": "^0.7.0",
 		"clsx": "^2.1.1",
-		"lucide-react": "^0.439.0",
-		"next": "14.2.9",
+		"lucide-react": "^0.441.0",
+		"next": "14.2.11",
 		"next-themes": "^0.3.0",
 		"react": "^18.3.1",
 		"react-dom": "^18.3.1",
@@ -31,14 +31,14 @@
 	"devDependencies": {
 		"@eslint/compat": "^1.1.1",
 		"@eslint/eslintrc": "^3.1.0",
-		"@next/eslint-plugin-next": "^14.2.9",
+		"@next/eslint-plugin-next": "^14.2.11",
 		"@ticketer/eslint-config": "workspace:*",
 		"@types/node": "^22.5.4",
 		"@types/react": "^18.3.5",
 		"@types/react-dom": "^18.3.0",
 		"autoprefixer": "^10.4.20",
 		"postcss": "^8.4.45",
-		"tailwindcss": "^3.4.10",
+		"tailwindcss": "^3.4.11",
 		"typescript": "^5.6.2"
 	}
 }
