+++ package.json
@@ -17,7 +17,7 @@
 		"eslint": "^9.10.0",
 		"prettier": "^3.3.3",
 		"prettier-plugin-tailwindcss": "^0.6.6",
-		"turbo": "^2.1.1"
+		"turbo": "^2.1.2"
 	},
 	"packageManager": "pnpm@9.7.0"
 }
