+++ packages/database/package.json
@@ -18,7 +18,7 @@
 		"@ticketer/env": "workspace:*",
 		"drizzle-orm": "^0.33.0",
 		"drizzle-zod": "^0.5.1",
-		"mysql2": "^3.11.0",
+		"mysql2": "^3.11.2",
 		"zod": "^3.23.8"
 	},
 	"devDependencies": {
