+++ packages/djs-framework/package.json
@@ -18,6 +18,6 @@
 	"devDependencies": {
 		"@ticketer/eslint-config": "workspace:*",
 		"@types/node": "^22.5.4",
-		"tsx": "^4.19.0"
+		"tsx": "^4.19.1"
 	}
 }
