+++ packages/eslint-config/package.json
@@ -8,12 +8,12 @@
 		"@eslint/eslintrc": "^3.1.0",
 		"@eslint/js": "^9.10.0",
 		"@types/eslint__js": "^8.42.3",
-		"eslint-config-next": "^14.2.9",
+		"eslint-config-next": "^14.2.11",
 		"eslint-config-prettier": "^9.1.0",
-		"eslint-config-turbo": "^2.1.1",
+		"eslint-config-turbo": "^2.1.2",
 		"eslint-plugin-drizzle": "^0.2.3",
 		"eslint-plugin-prettier": "^5.2.1",
-		"eslint-plugin-react": "^7.35.2",
+		"eslint-plugin-react": "^7.36.1",
 		"eslint-plugin-unicorn": "^55.0.0",
 		"typescript-eslint": "^8.5.0"
 	}
